# Batch model inputs to speed things up