## Background

Regardless of the task we're evaluating, we'll need to carry out the following steps:

* Preprocessing: generate language model (LM) inputs in whatever way is appropriate for the task
* Model inference: run the LM, possibly multiple times
* Postprocessing: take the LM's output, and score it for accuracy or the like

But if we just naively do that for each example, we run into a serious problem. If we run the examples one by one, we will drastically underutilize our GPUs, or whatever other accelerators we use.

To keep our GPUs busy, we need to process LM inputs in batches. Whatever batching system we go with, we will need to carry out the following tasks:
* Gathering LM requests and queueing them up for inference
* Taking a batch of requests and actually running the LM on them
* Following up on the LM's outputs

## This PR's changes

This PR introduces a `ModelRunner` that processes inputs in batches. With a reasonable batch size, we can go ~10x faster than without batching (batch size 1) 

Code | Model | Batch size | GPU | Wall-clock time (after model is loaded) | Accuracy on StoryCloze
-- | -- | -- | -- | -- | --
Before | GPT-2-117M | 1 | T4 | ~65 s | 52.4%
With PR | GPT-2-117M | 1 | V100 | 64 s | 52.4%
With PR | GPT-2-117M | 10 | V100 | 8.4 s | 52.4%
With PR | GPT-2-117M | 20 | V100 | 7.1 s | 52.4%
With PR | GPT-2-774M | 20 | V100 | 38.2 s | 66.1%